+++ KNXEngine/protocol/KNXHeader.js
@@ -20,15 +20,18 @@ class KNXHeader {
     }
     static createFromBuffer(buffer, offset = 0) {
         if (buffer.length < KNXConstants.KNX_CONSTANTS.HEADER_SIZE_10) {
+            if (sysLogger !== undefined && sysLogger !== null) sysLogger.error("KNXHeader: createFromBuffer: incomplete buffer. Buffer length: " + buffer.length + " expected HEADER_SIZE_10 equals to " + KNXConstants.KNX_CONSTANTS.HEADER_SIZE_10);
             throw new Error('Incomplete buffer');
         }
         const header_length = buffer.readUInt8(offset);
         if (header_length !== KNXConstants.KNX_CONSTANTS.HEADER_SIZE_10) {
+            if (sysLogger !== undefined && sysLogger !== null) sysLogger.error("KNXHeader: createFromBuffer: invalid header_length. header_length: " + header_length + " expected HEADER_SIZE_10 equals to " + KNXConstants.KNX_CONSTANTS.HEADER_SIZE_10);
             throw new Error(`Invalid buffer length ${header_length}`);
         }
         offset += 1;
         const version = buffer.readUInt8(offset);
         if (version !== KNXConstants.KNX_CONSTANTS.KNXNETIP_VERSION_10) {
+            if (sysLogger !== undefined && sysLogger !== null) sysLogger.error("KNXHeader: createFromBuffer: Unknown header version. Version: " + version + " expected KNXNETIP_VERSION_10 to " + KNXConstants.KNX_CONSTANTS.KNXNETIP_VERSION_10);
             throw new Error(`Unknown version ${version}`);
         }
         offset += 1;