+++ KNXEngine/protocol/KNXProtocol.js
@@ -17,6 +17,7 @@ const KNXTunnelingRequest = require("./KNXTunnelingRequest");
 const KNXTunnelingAck = require("./KNXTunnelingAck");
 const KNXRoutingIndication = require("./KNXRoutingIndication"); // 07/12/2021
 const KNXSecureSessionRequest = require("./KNXSecureSessionRequest"); // 07/12/2021
+const sysLogger = require("./../KnxLog.js").get(); // 08/04/2021 new logger to adhere to the loglevel selected in the config-window            
 
 const HPAI = require("./HPAI");